fix: move client tag and table setting btns to lower right
