+++ src/components/ConnectionTable.module.scss
@@ -83,10 +83,3 @@
 .rotate180 {
   transform: rotate(180deg);
 }
-
-.columnManagerRow {
-  width: 200px;
-  display: flex;
-  margin: 5px 0;
-  justify-content: space-between;
-}
