+++ src/components/Connections.module.scss
@@ -82,3 +82,10 @@
     width: 120px;
   }
 }
+
+.columnManagerRow {
+  width: 200px;
+  display: flex;
+  margin: 5px 0;
+  justify-content: space-between;
+}
